Moved request for courses to CourseForGroupController
