+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/group/StudentGroupController.java
@@ -5,11 +5,12 @@ import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import ua.edu.chdtu.deanoffice.api.group.dto.CourseForGroupDTO;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import ua.edu.chdtu.deanoffice.api.group.dto.StudentGroupDTO;
 import ua.edu.chdtu.deanoffice.api.group.dto.StudentGroupView;
-import ua.edu.chdtu.deanoffice.entity.CourseForGroup;
 import ua.edu.chdtu.deanoffice.entity.StudentGroup;
 import ua.edu.chdtu.deanoffice.service.CourseForGroupService;
 import ua.edu.chdtu.deanoffice.service.GroupService;
@@ -64,24 +65,5 @@ public class StudentGroupController {
     ) {
         List<StudentGroup> groups = groupService.getGroupsByDegreeAndYear(degreeId, year);
         return ResponseEntity.ok(parseToStudentGroupDTO(groups));
-    }
-
-    @GetMapping("{id}/courses")
-    @JsonView(StudentGroupView.Course.class)
-    public ResponseEntity getCourses(@PathVariable String id) {
-        List<CourseForGroup> courseForGroups = courseForGroupService.getCourseForGroup(Integer.parseInt(id));
-        return ResponseEntity.ok(parseToCourseForGroupDTO(courseForGroups));
-    }
-
-    private List<CourseForGroupDTO> parseToCourseForGroupDTO(List<CourseForGroup> courseForGroupList) {
-        Type listType = new TypeToken<List<CourseForGroupDTO>>() {}.getType();
-        return new ModelMapper().map(courseForGroupList, listType);
-    }
-
-    @GetMapping("{id}/{semester}/courses")
-    @JsonView(StudentGroupView.BasicCourse.class)
-    public ResponseEntity getCoursesBySemester(@PathVariable String id, @PathVariable String semester) {
-        List<CourseForGroup> courseForGroups = courseForGroupService.getCoursesForGroupBySemester(Integer.parseInt(id), Integer.parseInt(semester));
-        return ResponseEntity.ok(parseToCourseForGroupDTO(courseForGroups));
     }
 }
