+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/group/dto/StudentGroupView.java
@@ -3,10 +3,8 @@ package ua.edu.chdtu.deanoffice.api.group.dto;
 import ua.edu.chdtu.deanoffice.api.general.GeneralView;
 
 public class StudentGroupView {
-    public interface BasicGroup {}
+    public interface Basic {}
 
-    public interface WithStudents extends BasicGroup, GeneralView.PersonFullName {}
-    public interface GroupData extends BasicGroup {}
-    public interface BasicCourse extends GeneralView.Named, GeneralView.PersonFullName{}
-    public interface Course extends BasicCourse {}
+    public interface WithStudents extends Basic, GeneralView.PersonFullName {}
+    public interface GroupData extends Basic {}
 }
